+++ include/planners/drbt/DRGBT.h
@@ -29,10 +29,10 @@ namespace planning
             void generateHorizon();
             void updateHorizon(float d_c);
             void generateGBur();
-            void shortenHorizon(int num);
-            void addRandomStates(int num);
+            void shortenHorizon(size_t num);
+            void addRandomStates(size_t num);
             void addLateralStates();
-            bool modifyState(std::shared_ptr<planning::drbt::HorizonState> &q, int max_num_attempts);
+            bool modifyState(std::shared_ptr<planning::drbt::HorizonState> &q, size_t max_num_attempts);
             void computeReachedState(const std::shared_ptr<planning::drbt::HorizonState> q);
             void computeNextState();
             int getIndexInHorizon(const std::shared_ptr<planning::drbt::HorizonState> q);
@@ -44,8 +44,8 @@ namespace planning
             std::unique_ptr<planning::AbstractPlanner> initStaticPlanner(float max_planning_time);
             virtual void replan(float max_planning_time);
             void acquirePredefinedPath(const std::vector<std::shared_ptr<base::State>> &path_);
-            bool checkMotionValidity(int num_checks = DRGBTConfig::MAX_NUM_VALIDITY_CHECKS);
-            bool checkMotionValidity2(int num_checks = DRGBTConfig::MAX_NUM_VALIDITY_CHECKS);
+            bool checkMotionValidity(size_t num_checks = DRGBTConfig::MAX_NUM_VALIDITY_CHECKS);
+            bool checkMotionValidity2(size_t num_checks = DRGBTConfig::MAX_NUM_VALIDITY_CHECKS);
 
             std::vector<std::shared_ptr<planning::drbt::HorizonState>> horizon;     // List of all horizon states and their information
             std::shared_ptr<base::State> q_current;                                 // Current robot configuration
@@ -58,7 +58,7 @@ namespace planning
             bool replanning;                                                        // Whether path replanning is required
             base::State::Status status;                                             // The status of proceeding from 'q_current' towards 'q_next'
             std::vector<std::shared_ptr<base::State>> predefined_path;              // The predefined path that is being followed
-            int num_lateral_states;                                                 // Number of lateral states
+            size_t num_lateral_states;                                              // Number of lateral states
             float delta_q_max;                                                      // Maximal edge length when acquiring a new predefined path
             std::shared_ptr<planning::trajectory::Spline> spline_current;           // Current spline that 'q_current' is following in the current iteration
             std::shared_ptr<planning::trajectory::Spline> spline_next;              // Next spline that 'q_current' will follow until the end of current iteration
